Fixing memory leak.
This whole model is bad and I'll partially blame comments I read when starting out from the Agent class in aima-python where a point is made of "not letting the agent program see the rest of the agent."  This is unnecessary and misleading.  What else is there to see?  The model?  Well, then the agent is model-based and whether that model is hidden in a program variable or simply part of that particular agent's class is moot.  It does not cause issues in python (aside from many global functions floating about) but passing around these closures in strict Swift suggests a better design is in order.  The agent sees whatever it is designed to see.  Cf., Occam.