Finish EuclideanEnvironment protocol.
Sublest bug involved `mutating` modifier which compiler does not require on declaration, only on extension implementation, but which, when non-mutating step(Int:) tried to call mutating step(), the compiler could not even find the step() function! presumably because it is searching on exact signature so expected a non-mutating step() would be called from a non-mutating step(Int:).  So error message on step(Int:) was "you need to insert an Int when you call step();" i.e., it could not even find step().  Be cautious that compiler is using the function(s) you intend!  If signatures differ ever so slightly, the extension function will not look like it belongs to the protocol and will not be inherited by adopters!